Redone file architecture
