+++ Djipi/core/DjipiSystem.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <Log.h>
+
+namespace Djipi
+{
+	class DjipiSystem
+	{
+	public:
+		virtual ~DjipiSystem() {}
+
+	protected:
+		virtual bool Init() = 0;
+		virtual void Shutdown() = 0;
+	};
+}
